+++ pytorch_lightning/trainer/trainer.py
@@ -68,6 +68,7 @@ from pytorch_lightning.trainer.connectors.logger_connector import LoggerConnecto
 from pytorch_lightning.trainer.connectors.logger_connector.result import _ResultCollection
 from pytorch_lightning.trainer.connectors.signal_connector import SignalConnector
 from pytorch_lightning.trainer.data_loading import TrainerDataLoadingMixin
+from pytorch_lightning.trainer.optimizers import TrainerOptimizersMixin
 from pytorch_lightning.trainer.states import RunningStage, TrainerFn, TrainerState, TrainerStatus
 from pytorch_lightning.tuner.lr_finder import _LRFinder
 from pytorch_lightning.tuner.tuning import Tuner
@@ -117,6 +118,7 @@ warnings.filterwarnings(
 
 class Trainer(
     TrainerCallbackHookMixin,
+    TrainerOptimizersMixin,
     TrainerDataLoadingMixin,
 ):
     # Needed because of LightningOptimizer
